+++ CHANGELOG.md
@@ -7,12 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.16.0] - 2024-11-26
+
 ### Added
 
+- Add kubernetes events logging in Alloy.
 - Add support for Private CAs in alloy logs.
 - Add KubeEventsLogger option and related methods in loggedCLuster package.
 - Add `events-logger` flag in the operator.
 - Add toggle for `events-logger` in observability-bundle configmap.
+- Add tests for `alloy-events` in events-logger-config.
 
 ### Changed
 
@@ -402,7 +406,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add '--vintage' toggle
 - Add controller for Workload Management Cluster using cluster.x-k8s.io/v1beta1```
 
-[Unreleased]: https://github.com/giantswarm/logging-operator/compare/v0.15.2...HEAD
+[Unreleased]: https://github.com/giantswarm/logging-operator/compare/v0.16.0...HEAD
+[0.16.0]: https://github.com/giantswarm/logging-operator/compare/v0.15.2...v0.16.0
 [0.15.2]: https://github.com/giantswarm/logging-operator/compare/v0.15.1...v0.15.2
 [0.15.1]: https://github.com/giantswarm/logging-operator/compare/v0.15.0...v0.15.1
 [0.15.0]: https://github.com/giantswarm/logging-operator/compare/v0.14.0...v0.15.0